[modules] cleaned up opticalflow modules and edgeflow
