+++ conf/airframes/TUDELFT/tudelft_ardrone2_opticflow_edgeflow.xml
@@ -31,6 +31,8 @@
     <load name="cv_opticflow.xml">
       <define name="OPTICFLOW_DEBUG" value = "FALSE"/>
       <define name="OPTICFLOW_SHOW_FLOW" value = "FALSE"/>
+      <define name="USE_EDGEFLOW" value = "TRUE"/>
+      <define name="USE_LK" value = "FALSE"/>
     </load>
     <load name="video_thread.xml">
         <define name="VIDEO_THREAD_CAMERA" value="front_camera"/>
