+++ sw/airborne/modules/computer_vision/opticflow/opticflow_calculator.c
@@ -280,16 +280,19 @@ void opticflow_calc_frame(struct opticflow_t *opticflow, struct opticflow_state_
   // Velocity calculation
   // Right now this formula is under assumption that the flow only exist in the center axis of the camera.
   // TODO Calculate the velocity more sophisticated, taking into account the drone's angle and the slope of the ground plane.
-  float vel_hor = result->flow_der_x * result->fps * state->agl / opticflow->subpixel_factor  / OPTICFLOW_FX;
-  float vel_ver = result->flow_der_y * result->fps * state->agl / opticflow->subpixel_factor  / OPTICFLOW_FY;
+  float vel_x = result->flow_der_x * result->fps * state->agl / opticflow->subpixel_factor  / OPTICFLOW_FX;
+  float vel_y = result->flow_der_y * result->fps * state->agl / opticflow->subpixel_factor  / OPTICFLOW_FY;
+  result->vel_x = vel_x;
+  result->vel_y = vel_y;
 
   // Velocity calculation: uncomment if focal length of the camera is not known or incorrect.
   //  result->vel_x =  - result->flow_der_x * result->fps * state->agl / opticflow->subpixel_factor * OPTICFLOW_FOV_W / img->w
   //  result->vel_y =  result->flow_der_y * result->fps * state->agl / opticflow->subpixel_factor * OPTICFLOW_FOV_H / img->h
 
   // Rotate velocities from camera frame coordinates to body coordinates.
-  result->vel_x = vel_ver;
-  result->vel_y = - vel_hor;
+  // IMPORTANT for control! This the case on the ARDrone and bebop, but on other systems this might be different!
+  result->vel_body_x = vel_y;
+  result->vel_body_y = - vel_x;
 
   // Determine quality of noise measurement for state filter
   //TODO Experiment with multiple noise measurement models
